requirements.txt
fastapi
uvicorn
python-multipart
pdfplumber
httpx
python-dotenv
